refactor(cli): extract secondary CLI commands to commands.go
Move ~2,800 lines of secondary command constructors from main.go:
- newStopCmd, newStatusCmd, newInitCmd, newVersionCmd
- newTaskCmd with alerts/budget/quality integration
- newGitHubCmd and newGitHubRunCmd
- newBriefCmd
- newPatternsCmd + subcommands (list, search, stats, apply, ignore)
- newReplayCmd + subcommands (list, show, play, analyze, export, delete)
- newReleaseCmd, newAutopilotCmd + subcommands
- Helper functions: killExistingTelegramBot, checkForUpdates
- Dashboard mode: runDashboardMode, convertTaskStatesToDisplay
- Comment formatters: buildExecutionComment, buildFailureComment

This reduces main.go from ~6,249 to ~3,406 lines, improving
maintainability as part of the GH-1058 refactoring effort.

GH-1215
